refactor: split certify in several modules
