+++ crate/server/src/core/operations/validate.rs
@@ -19,11 +19,12 @@ use openssl::{
         store::X509StoreBuilder,
     },
 };
-use x509_parser::prelude::{FromDer, X509Certificate};
 
 use crate::{
     config::ProxyParams,
-    core::{KMS, retrieve_object_utils::retrieve_object_for_operation},
+    core::{
+        KMS, operations::certify::rfc9608, retrieve_object_utils::retrieve_object_for_operation,
+    },
     error::KmsError,
     result::KResult,
 };
@@ -607,23 +608,6 @@ async fn get_crl_bytes(
 ///
 /// # Errors
 ///
-/// Returns `true` if the certificate carries the `id-ce-noRevAvail` extension
-/// (OID 2.5.29.56, RFC 9608 §2), which signals that no revocation
-/// information is available. When present, CRL fetching is skipped for this cert.
-fn cert_has_no_rev_avail(cert: &X509) -> bool {
-    // OID 2.5.29.56 — id-ce-noRevAvail (RFC 9608 §2), { id-ce 56 }
-    // Encoded as DER value bytes (without tag/length): 55 1D 38
-    const NO_REV_AVAIL: &[u8] = &[0x55, 0x1d, 0x38];
-    let Ok(der) = cert.to_der() else { return false };
-    let Ok((_, parsed)) = X509Certificate::from_der(&der) else {
-        return false;
-    };
-    parsed
-        .extensions()
-        .iter()
-        .any(|ext| ext.oid.as_bytes() == NO_REV_AVAIL)
-}
-
 /// This function will return an error in the following cases:
 /// * If a certificate is found to be revoked or removed from the CRL.
 /// * If there is an issue deserializing a CRL.
@@ -661,7 +645,7 @@ pub(crate) async fn verify_crls(
         // RFC 9608 §4: skip CRL checks for certificates that carry id-ce-noRevAvail.
         // The extension signals that no revocation information is available; a
         // relying party MUST NOT reject the certificate for lack of a CRL/OCSP response.
-        if cert_has_no_rev_avail(certificate) {
+        if rfc9608::has_extension(certificate) {
             debug!("[{idx}] Certificate carries id-ce-noRevAvail (RFC 9608): skipping CRL check");
             continue;
         }
